at least works on some cases

+++ orderedhashmap.h
@@ -7,7 +7,7 @@
 typedef struct Node {
 	unsigned long key;
 	size_t value;
-	long long timestamp;
+	double timestamp;
 	struct Node *prev;
 	struct Node *next;
 } Node;
@@ -28,7 +28,7 @@ typedef struct {
 	size_t size;
 } LRUCache;
 
-long long get_first_node_time(LRUCache *cache);
+double get_first_node_time(LRUCache *cache);
 
 // Hash function for the keys
 unsigned long hash_function(unsigned long key);
